homework

+++ Lesson12_OOP/src/homework/Task02/Call.java
@@ -1,12 +1,14 @@
 package homework.Task02;
 
 public class Call {
+
     static double priceForAMinute = 0.11;
+
     private GSM caller;
     private GSM receiver;
     private double duration;
 
-    Call(GSM caller, GSM receiver, int duration) {
+    Call(GSM caller, GSM receiver, double duration) {
 	this.caller = caller;
 	this.receiver = receiver;
 	this.duration = duration;
